Create MapsUtilities to simplify MapsActivity
